review and pagination

+++ FirstAPI/Controllers/UserController.cs
@@ -3,11 +3,9 @@ using FirstAPI.Extensions;
 using FirstAPI.Services;
 using FirstAPI.ViewModels;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using SecureIdentity.Password;
-using System.Data;
 using System.Data.SqlClient;
 
 namespace FirstAPI.Controllers
@@ -26,19 +24,19 @@ namespace FirstAPI.Controllers
         [HttpPost("/users/register")]
         public async Task<IActionResult> Post([FromBody] RegisterViewModel model)
         {
-            if (!ModelState.IsValid)
-                return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
-
-            var parameters = new
+            try
             {
-                model.Name,
-                model.Email,
-                model.Password
-            };
+                if (!ModelState.IsValid)
+                    return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
 
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                try
+                var parameters = new
+                {
+                    model.Name,
+                    model.Email,
+                    Password = PasswordHasher.Hash(model.Password)
+                };
+
+                using (var connection = new SqlConnection(_connectionString))
                 {
                     const string sql = "INSERT INTO [Users] (Name, Email, Password) OUTPUT INSERTED.Id VALUES (@Name, @Email, @Password)";
 
@@ -55,15 +53,15 @@ namespace FirstAPI.Controllers
 
                     return Ok(new ResultViewModel<dynamic>(userData));
                 }
-                catch (SqlException ex) when (ex.Number == 2627)
-                {
-                    return StatusCode(400, new ResultViewModel<string>("05X99 - Este E-mail já está cadastrado"));
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
-                }
+            }
+            catch (SqlException ex) when (ex.Number == 2627)
+            {
+                return StatusCode(400, new ResultViewModel<string>("05X99 - Este E-mail já está cadastrado"));
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
             }
         }
 
@@ -72,30 +70,31 @@ namespace FirstAPI.Controllers
         [FromBody] LoginViewModel model,
         [FromServices] TokenService tokenService)
         {
-            if (!ModelState.IsValid)
-                return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
-
-            var parameters = new
+            try
             {
-                model.Email
-            };
+                if (!ModelState.IsValid)
+                    return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
 
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                const string sql = "SELECT * FROM [Users] WHERE Email = @Email";
-
-                await connection.OpenAsync();
-
-                var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
-
-                if (user == null)
-                    return StatusCode(401, new ResultViewModel<string>("Usuário não encontrado"));
-
-                if (model.Password != user.Password)
-                    return StatusCode(401, new ResultViewModel<string>("Usuário ou senha inválidos"));
-
-                try
+                var parameters = new
                 {
+                    model.Email
+                };
+
+                using (var connection = new SqlConnection(_connectionString))
+                {
+                    const string sql = "SELECT * FROM [Users] WHERE Email = @Email";
+
+                    await connection.OpenAsync();
+
+                    var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
+
+                    if (user == null)
+                        return StatusCode(401, new ResultViewModel<string>("Usuário não encontrado"));
+
+                    if (!PasswordHasher.Verify(user.Password, model.Password))
+                        return StatusCode(401, new ResultViewModel<string>("Usuário ou senha incorretos."));
+
+
                     var token = tokenService.GenerateToken(user);
 
                     var userData = new
@@ -107,11 +106,11 @@ namespace FirstAPI.Controllers
 
                     return Ok(new ResultViewModel<dynamic>(userData, null));
                 }
-                catch
-                {
-                    return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
+            }
+            catch
+            {
+                return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
 
-                }
             }
         }
 
@@ -121,6 +120,9 @@ namespace FirstAPI.Controllers
         {
             try
             {
+                if (!ModelState.IsValid)
+                    return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
+
                 var parameters = new { UserId = User.GetUserId() };
 
                 using (var connection = new SqlConnection(_connectionString))
@@ -154,82 +156,85 @@ namespace FirstAPI.Controllers
         [Authorize]
         public async Task<IActionResult> Put([FromBody] UpdateUserViewModel model)
         {
-            var parameters = new
+            try
             {
-                UserId = User.GetUserId(),
-                model.Name,
-                model.Email
-            };
+                if (!ModelState.IsValid)
+                    return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
 
-            using (var connection = new SqlConnection(_connectionString))
-            {
-                await connection.OpenAsync();
-
-                const string sql = "SELECT * FROM [Users] WHERE Id = @UserId";
-
-                var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
-
-                if (user == null)
-                    return NotFound(new ResultViewModel<string>("Usuário não encontrado"));
-
-                if (model.Password != user.Password)
-                    return StatusCode(401, new ResultViewModel<string>("Usuário ou senha inválidos"));
-
-                try
+                var parameters = new
                 {
+                    UserId = User.GetUserId(),
+                    model.Name,
+                    model.Email
+                };
+
+                using (var connection = new SqlConnection(_connectionString))
+                {
+                    await connection.OpenAsync();
+
+                    const string sql = "SELECT * FROM [Users] WHERE Id = @UserId";
+
+                    var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
+
+                    if (user == null)
+                        return NotFound(new ResultViewModel<string>("Usuário não encontrado"));
+
+                    if (!PasswordHasher.Verify(user.Password, model.Password))
+                        return StatusCode(401, new ResultViewModel<string>("Usuário ou senha incorretos."));
+
+
                     const string query = "UPDATE [Users] SET Name = @Name, Email = @Email WHERE Id = @UserId";
 
                     await connection.ExecuteAsync(query, parameters);
 
                     return Ok(new ResultViewModel<string>("Dados editados com sucesso!"));
                 }
-                catch (SqlException ex) when (ex.Number == 2627)
-                {
-                    return StatusCode(400, new ResultViewModel<string>("05X99 - Este E-mail já está cadastrado"));
-                }
-                catch (Exception ex)
-                {
-                    Console.WriteLine(ex.Message);
-                    return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
-                }
             }
+            catch (SqlException ex) when (ex.Number == 2627)
+            {
+                return StatusCode(400, new ResultViewModel<string>("05X99 - Este E-mail já está cadastrado"));
+            }
+            catch (Exception ex)
+            {
+                Console.WriteLine(ex.Message);
+                return StatusCode(500, new ResultViewModel<string>("05X04 - Falha interna no servidor"));
+            }
+
         }
 
         [HttpDelete("/users")]
         [Authorize]
         public async Task<IActionResult> Delete()
         {
-            var parameters = new { UserId = User.GetUserId() };
-
-            using (var connection = new SqlConnection(_connectionString))
+            try
             {
-                await connection.OpenAsync();
+                if (!ModelState.IsValid)
+                    return BadRequest(new ResultViewModel<string>(ModelState.GetErrors()));
 
-                const string sql = "SELECT * FROM [Users] WHERE Id = @UserId";
+                var parameters = new { UserId = User.GetUserId() };
 
-                var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
-
-                if (user == null)
-                    return NotFound(new ResultViewModel<string>("Usuário não encontrado"));
-
-                try
+                using (var connection = new SqlConnection(_connectionString))
                 {
+                    await connection.OpenAsync();
+
+                    const string sql = "SELECT * FROM [Users] WHERE Id = @UserId";
+
+                    var user = await connection.QueryFirstOrDefaultAsync<User>(sql, parameters);
+
+                    if (user == null)
+                        return NotFound(new ResultViewModel<string>("Usuário não encontrado"));
+
                     const string query = "DELETE FROM [Users] WHERE Id = @UserId";
 
                     await connection.ExecuteAsync(query, parameters);
 
                     return Ok(new ResultViewModel<string>("Conta cancelada com sucesso!"));
                 }
-                catch
-                {
-                    return StatusCode(500, new ResultViewModel<string>("Erro interno do servidor"));
-                }
+            }
+            catch
+            {
+                return StatusCode(500, new ResultViewModel<string>("Erro interno do servidor"));
             }
         }
-
-        //    var passwordHasher = new PasswordHasher<User>();
-        //    var result = passwordHasher.VerifyHashedPassword(user, user.Password, model.Password);
-        //    if (result != PasswordVerificationResult.Success)
-        //        return StatusCode(401, new ResultViewModel<string>("Usuário ou senha inválidos"));    
     }
 }
